refactor(header): :technologist: header component did not respect OCP and DIP SOLID principles, and thus SRP.
Logic related with language (such as working directly with localstorage and the whole logic) was moved into a single abstracted service.
